split auth and socketAuth, rename auth to authorize
