+++ packages/lib/server-side/src/index.ts
@@ -1,3 +1,5 @@
 export { Auth } from './auth';
+export { authorize } from './authorize';
 export { getCachedWebProvider } from './web3Provider/getCachedWebProvider';
 export type { IAccountDatabase } from './iAccountDatabase';
+export * from './utils';
