+++ packages/lib/server-side/src/utils.ts
@@ -1,93 +1,8 @@
 /* eslint-disable no-console */
-import { checkToken } from '@dm3-org/dm3-lib-delivery';
-import {
-    DeliveryServiceProfileKeys,
-    normalizeEnsName,
-} from '@dm3-org/dm3-lib-profile';
+import { DeliveryServiceProfileKeys } from '@dm3-org/dm3-lib-profile';
 import { ethers } from 'ethers';
 import { NextFunction, Request, Response } from 'express';
-import { Socket } from 'socket.io';
-import { ExtendedError } from 'socket.io/dist/namespace';
 import winston from 'winston';
-import type { IAccountDatabase } from './iAccountDatabase';
-
-export async function auth(
-    req: Request,
-    res: Response,
-    next: NextFunction,
-    ensName: string,
-    hasAccount: (ensName: string) => Promise<boolean>,
-    web3Provider: ethers.providers.JsonRpcProvider,
-    serverSecret: string,
-) {
-    const normalizedEnsName = normalizeEnsName(ensName);
-    const authHeader = req.headers['authorization'];
-    const token = authHeader && authHeader.split(' ')[1];
-    //TODO resolve addr for ens name
-    if (
-        token &&
-        (await checkToken(
-            web3Provider,
-            hasAccount,
-            normalizedEnsName,
-            token,
-            serverSecret,
-        ))
-    ) {
-        next();
-    } else {
-        console.warn('AUTH Token check failed for ', normalizedEnsName);
-        res.sendStatus(401);
-    }
-}
-
-export function socketAuth(
-    db: IAccountDatabase,
-    web3Provider: ethers.providers.JsonRpcProvider,
-    serverSecret: string,
-) {
-    return async (
-        socket: Socket,
-        next: (err?: ExtendedError | undefined) => void,
-    ) => {
-        try {
-            const ensName = normalizeEnsName(
-                socket.handshake.auth.account.ensName,
-            );
-            console.log('Start WS auth for ', ensName, socket.id);
-
-            if (
-                !(await checkToken(
-                    web3Provider,
-                    db.hasAccount,
-                    ensName,
-                    socket.handshake.auth.token as string,
-                    serverSecret,
-                ))
-            ) {
-                console.log('check token has failed for WS ');
-                return next(new Error('check token has failed for WS'));
-            }
-            const session = await db.getAccount(ensName);
-            if (!session) {
-                throw Error('Could not get session');
-            }
-            //we use session.account here as a key for setAccount here.
-            //We can do this because the address is used as account when the Session has been created.
-            //That saves a address lookup via ENS
-            await db.setAccount(session.account, {
-                ...session,
-                socketId: socket.id,
-            });
-        } catch (e) {
-            console.log('socket auth error');
-            console.log(e);
-            next(e as Error);
-        }
-
-        next();
-    };
-}
 
 export function logRequest(req: Request, res: Response, next: NextFunction) {
     winston.loggers.get('default').info({
